Implement operator precedence table
