--- puroboros/grammar
@@ -1,17 +1,0 @@
-expression additive_expression
-    ;
-
-additive_expression
-      multiplicative_expression
-    | additive_expression '+' multiplicative_expression
-    | additive_expression '-' multiplicative_expression
-    ;
-
-multiplicative_expression
-      number
-    | number '*' multiplicative_expression
-    | number '/' multiplicative_expression
-    ;
-
-number: T_INTLIT
-        ;
