+++ tests/test_expr.py
@@ -3,30 +3,45 @@ from unittest.mock import patch
 
 import pytest
 
-from puroboros.defs import ASTNode, ASTNodeType, Token, TokenType
 from puroboros.context import Context
+from puroboros.defs import ASTNode, ASTNodeType, Token, TokenType
 from puroboros.exceptions import ParserError
-from puroboros.expr import arithop, binexpr, primary
+from puroboros.expr import Parser
 from puroboros.scan import Scanner
 
 
+@pytest.fixture
+def parser():
+    context = Context()
+    scanner = Scanner(context)
+    parser = Parser(scanner)
+    return parser
+
+
+class TestParserSyntaxError:
+    def test_raises_syntax_error(self, parser):
+        parser.context.line = 10
+        with pytest.raises(ParserError) as e:
+            parser.raise_syntax_error()
+        
+        assert str(e.value) == 'Syntax error on line 10'
+
+
 class TestPrimary:
-    def test_primary_success(self):
-        context = Context()
+    def test_primary_success(self, parser):
         token = Token(TokenType.T_INTLIT, 5)
-        node = primary(token, context)
+        node = parser.primary(token)
 
         assert node.op == ASTNodeType.A_INTLIT
         assert node.intvalue == 5
         assert node.left is None
         assert node.right is None
 
-    def test_primary_failure(self):
-        context = Context()
+    def test_primary_failure(self, parser):
         token = Token(TokenType.T_STAR)
 
         with pytest.raises(ParserError) as e:
-            primary(token, context)
+            parser.primary(token)
 
         assert str(e.value) == 'Syntax error on line 1'
 
@@ -41,43 +56,39 @@ class TestArithmeticOperator:
             (TokenType.T_SLASH, ASTNodeType.A_DIVIDE),
         ]
     )
-    def test_arithop_success(self, token_type, expected_node_type):
-        node_type = arithop(token_type, Context())
+    def test_arithop_success(self, parser, token_type, expected_node_type):
+        node_type = parser.arith_op(token_type)
 
         assert node_type == expected_node_type
 
-    def test_arithop_failure(self):
+    def test_arithop_failure(self, parser):
         with pytest.raises(ParserError) as e:
-            arithop(TokenType.T_EOF, Context())
+            parser.arith_op(TokenType.T_EOF)
 
-        assert str(e.value) == 'Unknown token in arithop() on line 1'
+        assert str(e.value) == 'Syntax error on line 1'
 
 
 class TestBinaryExpression:
-    def test_one_level(self):
-        """
-        5
-        """
-        context = Context()
-        scanner = Scanner(context)
-        with patch.object(context, 'infile', StringIO('5')):
-            node = binexpr(scanner, context)
-
+    def test_int_literal(self, parser):
+        with patch.object(parser.context, 'infile', StringIO('1')):
+            node, op_token = parser.bin_expr()
+    
         assert node == ASTNode(
             op=ASTNodeType.A_INTLIT,
-            intvalue=5,
+            intvalue=1,
+        )
+        assert op_token == Token(
+            type=TokenType.T_EOF,
         )
 
-    def test_two_level(self):
+    def test_additive_expr(self, parser):
         """
           +
          / \
-        1   3
+        1   2
         """
-        context = Context()
-        scanner = Scanner(context)
-        with patch.object(context, 'infile', StringIO('1 + 3')):
-            node = binexpr(scanner, context)
+        with patch.object(parser.context, 'infile', StringIO('1 + 2')):
+            node, op_token = parser.bin_expr()
 
         assert node == ASTNode(
             op=ASTNodeType.A_ADD,
@@ -87,47 +98,129 @@ class TestBinaryExpression:
             ),
             right=ASTNode(
                 op=ASTNodeType.A_INTLIT,
+                intvalue=2,
+            ),
+        )
+        assert op_token == Token(
+            type=TokenType.T_EOF,
+        )
+
+    def test_double_additive_expr(self, parser):
+        """
+            +
+           / \
+          +   3
+         / \
+        1   2
+        """
+        with patch.object(parser.context, 'infile', StringIO('1 + 2 + 3')):
+            node, op_token = parser.bin_expr()
+        
+        assert node == ASTNode(
+            op=ASTNodeType.A_ADD,
+            left=ASTNode(
+                op=ASTNodeType.A_ADD,
+                left=ASTNode(
+                    op=ASTNodeType.A_INTLIT,
+                    intvalue=1,
+                ),
+                right=ASTNode(
+                    op=ASTNodeType.A_INTLIT,
+                    intvalue=2,
+                ),
+            ),
+            right=ASTNode(
+                op=ASTNodeType.A_INTLIT,
                 intvalue=3,
             ),
         )
+        assert op_token == Token(
+            type=TokenType.T_EOF,
+        )
 
-    def test_four_level(self):
+    def test_multiplicative_expr(self, parser):
         """
           *
          / \
-        2   +
-           / \
-          3   *
-             / \
-            4   5
+        1   2
         """
-        context = Context()
-        scanner = Scanner(context)
-        with patch.object(context, 'infile', StringIO('2 * 3 + 4 * 5')):
-            node = binexpr(scanner, context)
+        with patch.object(parser.context, 'infile', StringIO('1 * 2')):
+            node, op_token = parser.bin_expr()
 
         assert node == ASTNode(
             op=ASTNodeType.A_MULTIPLY,
             left=ASTNode(
                 op=ASTNodeType.A_INTLIT,
-                intvalue=2,
+                intvalue=1,
             ),
             right=ASTNode(
-                op=ASTNodeType.A_ADD,
+                op=ASTNodeType.A_INTLIT,
+                intvalue=2,
+            ),
+        )
+        assert op_token == Token(
+            type=TokenType.T_EOF,
+        )
+
+    def test_double_multiplicative_expr(self, parser):
+        """
+            *
+           / \
+          *   3
+         / \
+        1   2
+        """
+        with patch.object(parser.context, 'infile', StringIO('1 * 2 * 3')):
+             node, op_token = parser.bin_expr()
+
+        assert node == ASTNode(
+            op=ASTNodeType.A_MULTIPLY,
+            left=ASTNode(
+                op=ASTNodeType.A_MULTIPLY,
                 left=ASTNode(
                     op=ASTNodeType.A_INTLIT,
-                    intvalue=3,
+                    intvalue=1,
                 ),
                 right=ASTNode(
-                    op=ASTNodeType.A_MULTIPLY,
-                    left=ASTNode(
-                        op=ASTNodeType.A_INTLIT,
-                        intvalue=4,
-                    ),
-                    right=ASTNode(
-                        op=ASTNodeType.A_INTLIT,
-                        intvalue=5,
-                    ),
+                    op=ASTNodeType.A_INTLIT,
+                    intvalue=2,
                 ),
             ),
+            right=ASTNode(
+                op=ASTNodeType.A_INTLIT,
+                intvalue=3
+            ),
+        )
+        assert op_token == Token(
+            type=TokenType.T_EOF,
+        )
+
+    def test_mixed_expr(self, parser):
+        """
+            +
+           / \
+          *   3
+         / \
+        1   2
+        """
+        with patch.object(parser.context, 'infile', StringIO('1 * 2 + 3')):
+            node, op_token = parser.bin_expr()
+
+        assert node == ASTNode(
+            op=ASTNodeType.A_ADD,
+            left=ASTNode(
+                op=ASTNodeType.A_MULTIPLY,
+                left=ASTNode(
+                    op=ASTNodeType.A_INTLIT,
+                    intvalue=1,
+                ),
+                right=ASTNode(
+                    op=ASTNodeType.A_INTLIT,
+                    intvalue=2,
+                ),
+            ),
+            right=ASTNode(
+                op=ASTNodeType.A_INTLIT,
+                intvalue=3,
+            )
         )
